feat(rb-service): add executors with version (#74)
Using latest, an update of an image will
impact all services. In case of bug fixes,
it seems an advantage, but it may become
a nightmare in case of breaking changes.

That's why it is preferable to use executor
with image which target a specific version.

It allows to migrate easily and smoothly to
a new version of ruby or any other impactful
dependencies.

In order to keep these image up to date like
the latest image, we should refresh all tagged
images nightly.