fix(circleci/wait-job): fix edge case due to latency between builds
There is generally around 4 seconds between a job being successful and the following job being running.
It means it is possible that an older job similar to the current job just successful and that
the associate targeted job has not started yet.
In this case, two target jobs would run concurrently.
To avoid this edge case, we ensure that no jobs similar to the current job successful a few seconds ago.

[DVX-182]