Allow deploying latest build again (#114)
The current deployable check does not allow to redeploy the same
deployment once again. Meaning if your deployment gets corrupted
somehow you cannot delete it and deploy it again.

As an example, Airflow (jobflow) currently as a bug due to missing
support for certain kubernetes resources which can have it looses
track of some of the resources it created in k8s. The only work
around seem to delete and create again the deployment (_cf._
apache/airflow#21087).