feat(docker): ignore trivy in case of failure
It often occurs that Trivy fail blocking all image build and dev production.
If Trivy is needed in term of security, I think it does not bring enough value
to be a blocker.

This change will ignore trivy failure in case of exit code other than 0.
This default behavior will be overridable with `optional` parameter to false.